pyproject.toml
[tool.poetry]
name = "daytistics"
version = "0.1.0"
description = ""
authors = ["Leo Gall <gall.dev@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-allauth = "^64.0.0"
Django = "^5.0.7"
django-widget-tweaks = "^1.5.0"
pytest-django = "^4.8.0"
faker = "^26.1.0"
pytest-factoryboy = "^2.7.0"
django-mathfilters = "^1.0.0"
django-tailwind = "^3.8.0"
django-htmx = "^1.19.0"
django-browser-reload = "^1.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
